chore: use lombok for properties
