+++ simulator-spring-boot/src/test/java/org/citrusframework/simulator/model/ScenarioActionTest.java
@@ -1,27 +1,43 @@
+/*
+ * Copyright 2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.citrusframework.simulator.model;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class MessageTest {
+class ScenarioActionTest {
 
     @Test
     void equalsVerifier() throws Exception {
-        EntityTestUtils.equalsVerifier(Message.class);
+        EntityTestUtils.equalsVerifier(ScenarioAction.class);
 
-        Message message1 = new Message();
-        message1.setMessageId(1L);
+        ScenarioAction action1 = new ScenarioAction();
+        action1.setActionId(1L);
 
-        Message message2 = new Message();
-        message2.setMessageId(message1.getMessageId());
+        ScenarioAction action2 = new ScenarioAction();
+        action2.setActionId(action1.getActionId());
 
-        assertThat(message1).isEqualTo(message2);
+        assertThat(action1).isEqualTo(action2);
 
-        message2.setMessageId(2L);
-        assertThat(message1).isNotEqualTo(message2);
+        action2.setActionId(2L);
+        assertThat(action1).isNotEqualTo(action2);
 
-        message1.setMessageId(null);
-        assertThat(message1).isNotEqualTo(message2);
+        action1.setActionId(null);
+        assertThat(action1).isNotEqualTo(action2);
     }
 }
